requirements.txt
aiogram==2.22.1
Pillow==9.2.0

docx~=0.2.4

